Sync droid

+++ django/tpv_server/api_android/views/api_arqueos.py
@@ -5,13 +5,13 @@
 # @Last modified time: 2019-02-07T17:20:09+01:00
 # @License: Apache License v2.0
 
-from django.db.models import Q, Sum, F
+from django.db.models import Sum, F
 from tokenapi.http import JsonResponse
 from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 from django.db.models.fields import DecimalField
 from django.db import connection
-from api_android.tools import send_ticket_ws
+from api_android.tools import send_mensaje_impresora
 from api_android.tools.mails import send_cierre, getUsuariosMail
 from gestion.models import (Arqueocaja, Cierrecaja, Efectivo, Gastos,
                             Receptores, Ticket)
@@ -36,17 +36,9 @@ def get_cambio(request):
 
         return JsonResponse({"cambio":arqueo.cambio, "hay_arqueo": hay_arqueo})
     else:
-        return JsonResponse({"cambio":"0.00", "hay_arqueos": "True"})
+        return JsonResponse({"cambio":600.0, "hay_arqueo": True})
 
-@csrf_exempt
-def send_last_cierre(request):
-    arqueo = Arqueocaja.objects.all().order_by('-id').first()
 
-    users = getUsuariosMail()
-    for us in users:
-        send_cierre(us, arqueo.get_desglose_cierre())
-
-    return HttpResponse("success")
 
 @csrf_exempt
 def arquear(request):
@@ -63,10 +55,12 @@ def arquear(request):
             c.ticketfinal = ticketfinal.pk
         else:
             arqueo = Arqueocaja.objects.all().order_by("-id").first()
-            arqueo.gastos_set.all().delete()
-            arqueo.efectivo_set.all().delete()
-            arqueo.delete()
-            c = cierre
+            if arqueo:
+                arqueo.gastos_set.all().delete()
+                arqueo.efectivo_set.all().delete()
+                arqueo.delete()
+                c = cierre
+            
         crear_cierre(request, c)
     else:
         return HttpResponse("error")
@@ -130,45 +124,46 @@ def imprimir_desglose(request, arqueo):
     cambio = arqueo.cambio
     efectivo = arqueo.efectivo_set.all().values("moneda").annotate(can = Sum("can")).order_by("-moneda")
     retirar = arqueo.efectivo_set.all().aggregate(total=Sum(F("can") * F("moneda"), output_field=DecimalField()))['total']
-    retirar = float(retirar) - float(cambio)
-    lineas_retirada = []
-    lineas_cambio = []
-    parcial = 0
-    for linea in efectivo:
-        can = linea["can"]
-        moneda = linea["moneda"]
-        texto = "moneda" if moneda < 5 else "billete"
-        texto = texto + "s" if can > 1 else ""
-        if retirar <= parcial:
-            if can > 0:
-                lineas_cambio.append({"titulo": "Cambio", 'can':can,'tipo':float(moneda), 'texto_tipo': texto })
-        elif retirar > ((can * float(moneda)) + parcial):
-            parcial = parcial + float((can * moneda))
-            if can > 0:
-                lineas_retirada.append({"titulo": "Retirar", 'can':can,'tipo':float(moneda), 'texto_tipo': texto })
-        else:
-            diferencia = retirar - parcial
-            can_parcial = int(diferencia/float(moneda))
-            parcial = parcial + (can_parcial * float(moneda))
-            if can_parcial > 0:
-                lineas_retirada.append({"titulo": "Retirar", 'can':can_parcial,'tipo':float(moneda), 'texto_tipo': texto })
+    if retirar:
+        retirar = float(retirar) - float(cambio)
+        lineas_retirada = []
+        lineas_cambio = []
+        parcial = 0
+        for linea in efectivo:
+            can = linea["can"]
+            moneda = linea["moneda"]
             texto = "moneda" if moneda < 5 else "billete"
-            texto = texto + "s" if can_parcial > 1 else texto
-            if can - can_parcial > 0:
-                lineas_cambio.append({"titulo": "Cambio", 'can':can - can_parcial, 'tipo':float(moneda), 'texto_tipo': texto })
+            texto = texto + "s" if can > 1 else ""
+            if retirar <= parcial:
+                if can > 0:
+                    lineas_cambio.append({"titulo": "Cambio", 'can':can,'tipo':float(moneda), 'texto_tipo': texto })
+            elif retirar > ((can * float(moneda)) + parcial):
+                parcial = parcial + float((can * moneda))
+                if can > 0:
+                    lineas_retirada.append({"titulo": "Retirar", 'can':can,'tipo':float(moneda), 'texto_tipo': texto })
+            else:
+                diferencia = retirar - parcial
+                can_parcial = int(diferencia/float(moneda))
+                parcial = parcial + (can_parcial * float(moneda))
+                if can_parcial > 0:
+                    lineas_retirada.append({"titulo": "Retirar", 'can':can_parcial,'tipo':float(moneda), 'texto_tipo': texto })
+                texto = "moneda" if moneda < 5 else "billete"
+                texto = texto + "s" if can_parcial > 1 else texto
+                if can - can_parcial > 0:
+                    lineas_cambio.append({"titulo": "Cambio", 'can':can - can_parcial, 'tipo':float(moneda), 'texto_tipo': texto })
 
-    obj_cambio = {
-        "op": "desglose",
-        "receptor": Receptores.objects.get(nombre='Ticket').nomimp,
-        "fecha": datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f"),
-        "lineas": lineas_cambio
-    }
+        obj_cambio = {
+            "op": "desglose",
+            "receptor": Receptores.objects.get(nombre='Ticket').nomimp,
+            "fecha": datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f"),
+            "lineas": lineas_cambio
+        }
 
-    obj_desglose = {
-        "op": "desglose",
-        "receptor": Receptores.objects.get(nombre='Ticket').nomimp,
-        "fecha": datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f"),
-        "lineas": lineas_retirada
-    }
-    send_ticket_ws(request, obj_desglose)
-    send_ticket_ws(request, obj_cambio)
+        obj_desglose = {
+            "op": "desglose",
+            "receptor": Receptores.objects.get(nombre='Ticket').nomimp,
+            "fecha": datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f"),
+            "lineas": lineas_retirada
+        }
+        send_mensaje_impresora(obj_desglose)
+        send_mensaje_impresora(obj_cambio)
